requirements.txt
flask==2.2.3
python-dotenv==1.0.0
pyodbc==4.0.39
azure-storage-blob==12.16.0
flask-wtf==1.1.1
pillow==9.5.0